Prototype 1 - WORKING
after working hours and hours to get this working i’ve finally managed
to get it colliding with the triangles, squares and circles.  It
collides with flippers and absorbers too but to get it working for the
prototype I just commented them out. here’s my code for the prototype
that we’re submitting.